moved TerrainData to package mapgenerators.data
