+++ src/main/kotlin/de/gleex/pltcmd/model/terrain/Terrain.kt
@@ -1,6 +1,6 @@
 package de.gleex.pltcmd.model.terrain
 
-import de.gleex.pltcmd.model.mapgenerators.TerrainData
+import de.gleex.pltcmd.model.mapgenerators.data.TerrainData
 import kotlin.random.Random
 
 data class Terrain private constructor(val type: TerrainType, val height: TerrainHeight) {
